统一PC与手机的发布字段（必填信息与选填信息）

+++ qingdaofu/TabBarController/消息/controller/发布消息/PublishMessagesViewController.m
@@ -200,7 +200,7 @@
 - (void)footerRefreshWithMessageOfPublish
 {
     _mesPage++;
-    NSString *page = [NSString stringWithFormat:@"%d",_mesPage];
+    NSString *page = [NSString stringWithFormat:@"%ld",(long)_mesPage];
     [self getPublishMessageListWithPage:page];
     dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(0.5 * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
         [self.newsListTableView footerEndRefreshing];
